new controller, gateway, etc + reformatted tests + loosely coupled structure
