+++ CheckYourEligibility.API.Tests/Controllers/GetAllBulkChecksControllerTests.cs
@@ -39,7 +39,7 @@ namespace CheckYourEligibility.API.Tests.Controllers
                     _mockLogger.Object,
                 mockAudit.Object,
                 configuration,
-                    null!, null!, null!, null!, null!, null!, null!,
+                    null!, null!, null!, null!, null!,
                     _mockUseCase.Object
                 );
