+++ CheckYourEligibility.API/Controllers/BulkCheckController.cs
@@ -28,8 +28,6 @@ public class BulkCheckController : BaseController
     private readonly IGetBulkUploadResultsUseCase _getBulkUploadResultsUseCase;
     private readonly IDeleteBulkCheckUseCase _deleteBulkUploadUseCase;
     private readonly IGetAllBulkChecksUseCase _getAllBulkChecksUseCase;
-    private readonly IEligibilityCheckReportUseCase _eligibilityCheckReportUseCase;
-    private readonly IGetEligibilityReportHistoryUseCase _getEligibilityReportHistoryUseCase;
     private readonly ILogger<BulkCheckController> _logger;
     private readonly string _localAuthorityScopeName;
 
@@ -42,8 +40,6 @@ public class BulkCheckController : BaseController
         IGetBulkUploadProgressUseCase getBulkUploadProgressUseCase,
         IGetBulkUploadResultsUseCase getBulkUploadResultsUseCase,
         IDeleteBulkCheckUseCase deleteBulkUploadUseCase,
-        IEligibilityCheckReportUseCase eligibilityCheckReportUseCase,
-        IGetEligibilityReportHistoryUseCase getEligibilityReportHistoryUseCase,
         IGetAllBulkChecksUseCase getAllBulkChecksUseCase
     )
         : base(audit)
@@ -57,8 +53,6 @@ public class BulkCheckController : BaseController
         _getBulkUploadProgressUseCase = getBulkUploadProgressUseCase;
         _getBulkUploadResultsUseCase = getBulkUploadResultsUseCase;
         _deleteBulkUploadUseCase = deleteBulkUploadUseCase;
-        _eligibilityCheckReportUseCase = eligibilityCheckReportUseCase;
-        _getEligibilityReportHistoryUseCase = getEligibilityReportHistoryUseCase;
         _getAllBulkChecksUseCase = getAllBulkChecksUseCase;
     }
 
@@ -504,99 +498,5 @@ public class BulkCheckController : BaseController
             return StatusCode(StatusCodes.Status403Forbidden, new ErrorResponse { Errors = [new Error { Title = ex.Message }] });
         }
     }
-
-    /// <summary>
-    ///     Returns reports of asscioated bulk checks between a set time period
-    /// </summary>
-    /// <returns></returns>
-    [ProducesResponseType(typeof(EligibilityCheckReportResponse), (int)HttpStatusCode.OK)]
-    [ProducesResponseType(typeof(ErrorResponse), (int)HttpStatusCode.BadRequest)]
-    [Consumes("application/json", "application/vnd.api+json;version=1.0")]
-    [HttpPost("/check-eligibility/report")]
-    [Authorize(Policy = PolicyNames.RequireBulkCheckScope)]
-    [Authorize(Policy = PolicyNames.RequireLaOrMatOrSchoolScope)]
-    public async Task<ActionResult> EligibilityCheckReportRequest([FromBody] EligibilityCheckReportRequest model)
-    {
-        try
-        {
-            var localAuthorityIds = User.GetSpecificScopeIds(_localAuthorityScopeName);
-            if (localAuthorityIds == null || localAuthorityIds.Count == 0)
-            {
-                return BadRequest(new ErrorResponse
-                {
-                    Errors = [new Error { Title = "No local authority scope found" }]
-                });
-            }
-
-            var result = await _eligibilityCheckReportUseCase.Execute(model);
-
-            return new ObjectResult(result) { StatusCode = StatusCodes.Status200OK };
-        }
-        catch (UnauthorizedAccessException ex)
-        {
-            return Unauthorized(new ErrorResponse
-            {
-                Errors = [new Error { Title = ex.Message }]
-            });
-        }
-        catch (NotFoundException)
-        {
-            return NotFound(new ErrorResponse
-            { Errors = [new Error { Status = StatusCodes.Status404NotFound }] });
-        }
-        catch (FluentValidation.ValidationException ex)
-        {
-            return BadRequest(new ErrorResponse { Errors = [new Error { Title = ex.Message }] });
-        }
-        catch (ValidationException ex)
-        {
-            return BadRequest(new ErrorResponse { Errors = ex.Errors });
-        }
-    }
-
-    /// <summary>
-    ///     Gets all report history for a given LA
-    /// </summary>
-    /// <returns></returns>
-    [ProducesResponseType(typeof(EligibilityCheckReportHistoryResponse), (int)HttpStatusCode.OK)]
-    [ProducesResponseType(typeof(ErrorResponse), (int)HttpStatusCode.BadRequest)]
-    [Consumes("application/json", "application/vnd.api+json;version=1.0")]
-    [HttpGet("/check-eligibility/report-history/{localAuthorityId}")]
-    [Authorize(Policy = PolicyNames.RequireBulkCheckScope)]
-    [Authorize(Policy = PolicyNames.RequireLaOrMatOrSchoolScope)]
-    public async Task<ActionResult> GetAllReportHistory(string localAuthorityId)
-    {
-        try
-        {
-            var localAuthorityIds = User.GetSpecificScopeIds(_localAuthorityScopeName);
-            if (localAuthorityIds == null || localAuthorityIds.Count == 0)
-            {
-                return BadRequest(new ErrorResponse
-                {
-                    Errors = [new Error { Title = "No local authority scope found" }]
-                });
-            }
-
-            var result = await _getEligibilityReportHistoryUseCase.Execute(localAuthorityId, localAuthorityIds);
-
-            return new ObjectResult(result) { StatusCode = StatusCodes.Status200OK };
-        }
-        catch (UnauthorizedAccessException ex)
-        {
-            return Unauthorized(new ErrorResponse
-            {
-                Errors = [new Error { Title = ex.Message }]
-            });
-        }
-        catch (FluentValidation.ValidationException ex)
-        {
-            return BadRequest(new ErrorResponse { Errors = [new Error { Title = ex.Message }] });
-        }
-        catch (ValidationException ex)
-        {
-            return BadRequest(new ErrorResponse { Errors = ex.Errors });
-        }
-    }
-
 
 }
