+++ CheckYourEligibility.API/Gateways/Interfaces/IEligibilityCheckReporting.cs
@@ -1,0 +1,5 @@
+public interface IEligibilityCheckReporting
+{
+    Task<IEnumerable<EligibilityCheckReportResponseItem>> EligibilityCheckReports(EligibilityCheckReportRequest request, CancellationToken cancellationToken = default);
+    Task<IEnumerable<EligibilityCheckReportHistoryItem>> GetEligibilityCheckReportHistory(string localAuthorityId); 
+}
