+++ tests/cypress/API/08 - EligibilityCheckReporting/GetEligibilityReportHistory.cy.ts
